fix: Attach compensation inner elements in a sequence
